requirements.txt
backports.functools_lru_cache ; python_version < '2.9'